Рекурсивная реализация с использованием стека на куче

+++ Hash_func/md5.cpp
@@ -340,7 +340,7 @@ std::string MD5::hexdigest() const
 		return "";
 
 	char buf[33];
-	for (int i = 0; i < 16; i++)
+	for (unsigned long long i = 0; i < 16; i++)
 		sprintf(buf + i * 2, "%02x", digest[i]);
 	buf[32] = 0;
